Using 'cin'
To prevent any unexpected behavior while using cin commands in C++, it's important to clear the buffer space to ensure that there is no residual data left in the input buffer. This can be achieved by using the ignore() function or by using cin.ignore() along with cin.clear() commands.