Calculate area of triangle using herons formula
Here is a C++ program that can calculate the area of any triangle using Heron's formula. The program requires the user to input the length of the three sides of the triangle.

Once the user has entered the values, the program will calculate the area of the triangle using the formula.